+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 ### Breaking Changes
 
 ### Added
+
+## 4.5.0 - 2024-02-06
+
+### Added
 - Support for running tasks against individual databases in a multi-database setup with Rails >= 6.1 ([#930](https://github.com/grosser/parallel_tests/pull/930))
 
 ## 4.4.0 - 2023-12-24